Move multi-step plugin to /plugins
