+++ src/server/apply-changelog/controllers/change-details-form.js
@@ -5,7 +5,7 @@ import { buildOptions } from '../../common/helpers/options/build-options.js'
 import { getEnvironments } from '../../common/helpers/environments/get-environments.js'
 import { getAdditionalData } from '../helpers/get-additional-data.js'
 import { detailsValidation } from '../helpers/schema/details-validation.js'
-import { provideStepData } from '../../common/helpers/multistep-form/provide-step-data.js'
+import { provideStepData } from '#server/plugins/multistep-form/provide-step-data.js'
 import { fetchPostgresServices } from '../helpers/fetchers.js'
 import { fetchLatestMigrations } from '../../common/helpers/fetch/fetch-latest-migrations.js'
 import { provideDatabaseStatusClassname } from '../../common/components/database-detail/provide-database-status-classname.js'
