+++ src/server/deploy-service/controllers/deploy/details-form.js
@@ -5,7 +5,7 @@ import { buildOptions } from '../../../common/helpers/options/build-options.js'
 import { getEnvironments } from '../../../common/helpers/environments/get-environments.js'
 import { getAdditionalData } from '../../helpers/get-additional-data.js'
 import { detailsValidation } from '../../helpers/schema/details-validation.js'
-import { provideStepData } from '../../../common/helpers/multistep-form/provide-step-data.js'
+import { provideStepData } from '../../../plugins/multistep-form/provide-step-data.js'
 import { fetchLatestMigrations } from '../../../common/helpers/fetch/fetch-latest-migrations.js'
 import { provideDatabaseStatusClassname } from '../../../common/components/database-detail/provide-database-status-classname.js'
 import { nullify404 } from '../../../common/helpers/nullify-404.js'
