+++ src/server/deploy-service/controllers/deploy/options-form.js
@@ -6,7 +6,7 @@ import { buildOptions } from '../../../common/helpers/options/build-options.js'
 import { availableInstances } from '../../constants/available-instances.js'
 import { provideFormValues } from '../../helpers/pre/provide-form-values.js'
 import { cpuToVCpu } from '../../helpers/cpu-to-vcpu.js'
-import { provideStepData } from '../../../common/helpers/multistep-form/provide-step-data.js'
+import { provideStepData } from '../../../plugins/multistep-form/provide-step-data.js'
 import { transformRunningServices } from '../../../services/service/about/transformers/running-services.js'
 import { getEnvironments } from '../../../common/helpers/environments/get-environments.js'
 import { provideDatabaseStatusClassname } from '../../../common/components/database-detail/provide-database-status-classname.js'
