+++ src/server/deploy-service/routes.js
@@ -7,7 +7,7 @@ import { detailsController } from './controllers/deploy/details.js'
 import { detailsFormController } from './controllers/deploy/details-form.js'
 import { summaryController } from './controllers/deploy/summary.js'
 import { deployController } from './controllers/deploy/deploy.js'
-import { multistepForm } from '../common/helpers/multistep-form/multistep-form.js'
+import { multistepForm } from '../plugins/multistep-form/multistep-form.js'
 import { availableVersionsController } from './controllers/available-versions.js'
 import { availableMemoryController } from './controllers/available-memory.js'
 import { urlTemplates, formSteps } from './helpers/multistep-form/steps.js'
